requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
oauthlib==3.2.2
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
